==> 00_data_preperation.ipynb <==
# default_exp data_preperation

# Data preperation

> Clean Swissmetro dataset.

#hide
from nbdev.showdoc import *

import pandas as pd
from sklearn.model_selection import train_test_split

### Data preperation

df = pd.read_csv('./data/swissmetro.dat', sep='	')
# Remove unknown age (6)
df['AGE'].value_counts
df = df[df['AGE'] != 6]
# Remove "other" trip purpose (9)
df = df[df['PURPOSE'] != 9]
# Remove "unknown" choice (0)
df = df[df['CHOICE'] != 0]
df.shape

# Save to csv file
df.to_csv('./data/swissmetro_clean.csv',index=False)

### Data exploration

# Reload data
data = pd.read_csv('./data/swissmetro_clean.csv')
data.shape

data.head()

data.columns

### Training, Validation and Test

# Divide the dataset into training (70%), validation (15%) and test (15%), same as in TasteNet-MNL models
train, test = train_test_split(data, test_size=0.3, shuffle=True, stratify=data.CHOICE.values)
valid, test = train_test_split(test, test_size=0.5, shuffle=True, stratify=test.CHOICE.values)
train.shape, valid.shape, test.shape

# Check percentage of CHOICE in train, valid, and test sets
train.CHOICE.value_counts()

valid.CHOICE.value_counts()

test.CHOICE.value_counts()

# Save into csv files
train.to_csv('./data/swissmetro_train.csv', index=False)
valid.to_csv('./data/swissmetro_valid.csv', index=False)
test.to_csv('./data/swissmetro_test.csv', index=False)
