==> 01_larch_mnl.ipynb <==
import larch, numpy, pandas, os
from larch import P, X

data = pandas.read_csv("./data/swissmetro.dat", sep="\t")
condition = (data.PURPOSE.isin([1,3]) & (data.CHOICE != 0))
df = data.loc[condition].copy()
df.shape

df.columns

# Scale travel time and travel cost
df[['TRAIN_TT', 'SM_TT', 'CAR_TT']] = df[['TRAIN_TT', 'SM_TT', 'CAR_TT']] / 100
df["Free_ticket"] = 1 - df.GA
df['TRAIN_CO'] = df['TRAIN_CO'] * df["Free_ticket"] / 100
df['SM_CO'] = df['SM_CO'] * df["Free_ticket"] / 100
df['CAR_CO'] = df['CAR_CO'] * df["Free_ticket"] / 100

# For clarity, define these mode
TRAIN = 1
SM = 2
CAR = 3

alt_codes = [TRAIN, SM, CAR]
alt_names = ["Train", "SwissMetro", "Car"]

dfs = larch.DataFrames(co=df, alt_codes=alt_codes, 
                       alt_names=alt_names, ch_name="CHOICE")

m = larch.Model(dataservice=dfs)
m.title = "Travel mode choice"

df.head().T

m.utility_co[TRAIN] = (
    + P.ASC_TRAIN
    + P.TT * X.TRAIN_TT
    + P.CO * X.TRAIN_CO
)
m.utility_co[SM] = (
    + P.TT * X.SM_TT
    + P.CO * X.SM_CO
)
m.utility_co[CAR] = (
    + P.ASC_CAR
    + P.TT * X.CAR_TT
    + P.CO * X.CAR_CO
)

m.choice_co_code = "CHOICE"

m.availability_co_vars = {
    TRAIN: 'TRAIN_AV',
    SM: 'SM_AV',
    CAR: 'CAR_AV',
}

### Model Estimation

m.load_data()

m.dataframes.choice_avail_summary()

result = m.maximize_loglike(method='slsqp')

m.parameter_summary()

#### NOTE:
# The results seem different from Biogeme and Pylogit.
# This may need to be reinvestigated again !!!

 dir(P)

P??
