==> 01_pylogit_data.ipynb <==
# Source code from: https://github.com/timothyb0912/pylogit

from collections import OrderedDict
import numpy as np
import pandas as pd
import pylogit as pl

### Load and clean Swiss metro data

sm_wide = pd.read_csv("./data/swissmetro.dat", sep="\t")
condition = (sm_wide.PURPOSE.isin([1,3]) & (sm_wide.CHOICE != 0))
sm_wide = sm_wide.loc[condition].copy()
sm_wide.shape

sm_wide.head(10).T

### Convert Wide to Long format

choice = "CHOICE"
# Define alternative variables
alt_vars = {u'travel_time': dict([(1, 'TRAIN_TT'),
                                   (2, 'SM_TT'),
                                   (3, 'CAR_TT')]),
              u'travel_cost': dict([(1, 'TRAIN_CO'),
                                    (2, 'SM_CO'),
                                    (3, 'CAR_CO')]),
              u'headway': dict([(1, 'TRAIN_HE'),
                                (2, 'SM_HE')]),
              u'seat_configuration': dict([(2, "SM_SEATS")])}

avai_vars = {1: 'TRAIN_AV',
              2: 'SM_AV', 
              3: 'CAR_AV'}

def wide2long(df_wide, choice_col, alternative_vars, available_vars):
    """
    This function convert wide format data frame to long format data frame
    :param df_wide: Pandas Data frame in wide format
    :param choice_col: str, name of choice column
    :param alternative_vars: a dictionary of alternative variables
    :param available_vars: a dictionary of available variables
    :return df_long: Pandas Data frame in long format
    """
    # Get individual variables
    no_ind_vars = not_ind_vars(choice_col, alternative_vars, available_vars)
    ind_vars = list(set(sm_wide.columns) - set(no_ind_vars))
    # Generate custom_id column
    df_wide["custom_id"] =  np.arange(df_wide.shape[0], dtype=int) + 1
    # Convert wide to long using pylogit function
    df_long = pl.convert_wide_to_long(df_wide, ind_vars, alternative_vars, available_vars,
                                      "custom_id", choice_col, "mode_id")    
    return df_long

def not_ind_vars(choice_col, alternative_vars, available_vars):
    not_ind_vars = [choice_col]
    for d in alternative_vars.values():
        not_ind_vars.extend(list(d.values()))
    not_ind_vars.extend(list(avai_vars.values()))
    return not_ind_vars

# Convert wide to long format
sm_long = wide2long(sm_wide, choice, alt_vars, avai_vars)

sm_long.shape

sm_long.head(10).T

### Save long format

sm_long.to_csv("./data/swissmetro_long.csv", index=False)
