==> 01_pylogit_mnl.ipynb <==
# Source code from: https://github.com/timothyb0912/pylogit

from collections import OrderedDict
import numpy as np
import pandas as pd
import pylogit as pl

### Load Swiss metro data

sm_long = pd.read_csv("./data/swissmetro_long.csv")
sm_long.shape

### Create the model specification
The model specification being used in this example is the following:
$$
\begin{aligned}
V_{i, \textrm{Train}} &= \textrm{ASC Train} + \\
&\quad \beta _{ \textrm{tt_train} } \textrm{Travel Time} _{ \textrm{Train}} + \\
&\quad \beta _{ \textrm{tc_train} } \textrm{Travel Cost}_{\textrm{Train}} * \left( GA == 0 \right) + \\
&\quad \beta _{ \textrm{headway_train} } \textrm{Headway} _{\textrm{Train}} \\
\\
V_{i, \textrm{Swissmetro}} &= \textrm{ASC Swissmetro} + \\
&\quad \beta _{ \textrm{tt_sm} } \textrm{Travel Time} _{ \textrm{Swissmetro}} + \\
&\quad \beta _{ \textrm{tc_sm} } \textrm{Travel Cost}_{\textrm{Swissmetro}} * \left( GA == 0 \right) + \\
&\quad \beta _{ \textrm{headway_sm} } \textrm{Heaway} _{\textrm{Swissmetro}} \\
% &\quad \beta _{ \textrm{seat} } \left( \textrm{Seat Configuration} == 1 \right) \\
% &\quad \beta _{ \textrm{survey} } \left( \textrm{Train Survey} == 1 \right) \\
% &\quad \beta _{ \textrm{first_class} } \left( \textrm{First Class} == 0 \right) \\
\\
V_{i, \textrm{Car}} &= \beta _{ \textrm{tt_car} } \textrm{Travel Time} _{ \textrm{Car}} + \\
&\quad \beta _{ \textrm{tc_car}} \textrm{Travel Cost}_{\textrm{Car}} \\
% &\quad \beta _{\textrm{luggage}=1} \left( \textrm{Luggage} == 1 \right) + \\
% &\quad \beta _{\textrm{luggage}>1} \left( \textrm{Luggage} > 1 \right)
\end{aligned}
$$

# Update travel time (minutes => hours)
# sm_long["travel_time"] = sm_long["travel_time"] 
# Get free_ticket
sm_long["free_ticket"] = (((sm_long["GA"] == 1) | (sm_long["WHO"]/ 100.0 == 2)) &
                            sm_long["mode_id"].isin([1,2])).astype(int)
# Update travel cost
sm_long["travel_cost"] = (sm_long["travel_cost"] * (sm_long["free_ticket"] == 0)) 

spec = OrderedDict()
spec["intercept"] = [1, 2]
spec["travel_time"] = [1, 2, 3]
spec["travel_cost"] = [1, 2, 3]
spec["headway"] = [1, 2]

# Estimate the multinomial logit model (MNL)
sm_mnl = pl.create_choice_model(data=sm_long,
                                alt_id_col="mode_id",
                                obs_id_col="custom_id",
                                choice_col="CHOICE",
                                specification=spec,
                                model_type="MNL")

# Specify the initial values and method for the optimization.
sm_mnl.fit_mle(np.zeros(10))

# Look at the estimation results
sm_mnl.get_statsmodels_summary()

### Evaluation metrics

from sklearn.metrics import accuracy_score, f1_score, classification_report

def model_pred(data, model, alt_id_col, obs_id_col, choice_col):
    data['predicted'] = model.predict(data)
    is_chosen = data.groupby([obs_id_col])['predicted'].idxmax()
    data['predicted_choice'] = 0
    data.loc[is_chosen.values,'predicted_choice'] = 1
    
    actual = data.loc[data[choice_column] == 1, alt_id_column]
    pred = data.loc[data['predicted_choice'] == 1, alt_id_column]
    return actual, pred   

def nll():
    pass

alt_id_column = "mode_id"
obs_id_column = "custom_id"
choice_column = "CHOICE"

##### MNL model results

actual, predict = model_pred(sm_long, sm_mnl, alt_id_column, obs_id_column, choice_column)
print(classification_report(actual, predict))

acc = accuracy_score(actual, predict), 
f1 = f1_score(actual, predict, average="weighted")
print(f"Accuracy = {np.round(acc,5)}, F1_score = {np.round(f1,5)}")

########################################################################################

## Simple MNL as in 1.logit Biogeme

The model specification being used in this example is the following:
$$
\begin{aligned}
V_{i, \textrm{Train}} &= \textrm{ASC Train}  + \\
&\quad \beta _{ \textrm{tt} } \textrm{Travel Time} _{ \textrm{Train}} + \\
&\quad \beta _{ \textrm{tc} } \textrm{Travel Cost}_{\textrm{Train}} * \left( GA == 0 \right) + \\
\\
V_{i, \textrm{Swissmetro}} &= \quad \beta _{ \textrm{tt} } \textrm{Travel Time} _{ \textrm{Swissmetro}} + \\
&\quad \beta _{ \textrm{tc} } \textrm{Travel Cost}_{\textrm{Swissmetro}} * \left( GA == 0 \right) + \\
\\
V_{i, \textrm{Car}} &= \textrm{ASC Car} + \\
&\beta _{ \textrm{tt} } \textrm{Travel Time} _{ \textrm{Car}} + \\
&\quad \beta _{ \textrm{tc}} \textrm{Travel Cost}_{\textrm{Car}} \\
\end{aligned}
$$

# Read data
sm_long = pd.read_csv("./data/swissmetro_long.csv")

# Update travel time (minutes => hours)
sm_long["travel_time"] = sm_long["travel_time"] / 100
# Get free_ticket
sm_long["free_ticket"] = (((sm_long["GA"] == 1) | (sm_long["WHO"]/ 100.0 == 2)) &
                            sm_long["mode_id"].isin([1,2])).astype(int)
# Update travel cost
sm_long["travel_cost"] = (sm_long["travel_cost"] * (sm_long["free_ticket"] == 0)) / 100

spec = OrderedDict()
spec["intercept"] = [3, 1]
spec["travel_cost"] = [[1, 2, 3]]
spec["travel_time"] = [[1, 2, 3]]

# Estimate the multinomial logit model (MNL)
sm_mnl = pl.create_choice_model(data=sm_long,
                                alt_id_col="mode_id",
                                obs_id_col="custom_id",
                                choice_col="CHOICE",
                                specification=spec,
                                model_type="MNL")

# Specify the initial values and method for the optimization.
sm_mnl.fit_mle(np.zeros(4))

# Look at the estimation results
sm_mnl.get_statsmodels_summary()

actual, predict = model_pred(sm_long, sm_mnl, alt_id_column, obs_id_column, choice_column)
print(classification_report(actual, predict))

acc = accuracy_score(actual, predict), 
f1 = f1_score(actual, predict, average="weighted")
print(f"Accuracy = {np.round(acc,5)}, F1_score = {np.round(f1,5)}")
