==> 01_pylogit_mnl_complex.ipynb <==
# Source code from: https://github.com/timothyb0912/pylogit

from collections import OrderedDict
import numpy as np
import pandas as pd
import pylogit as pl

### Load Swiss metro data

sm_long = pd.read_csv("./data/swissmetro_long.csv")
sm_long.shape

### Create the model specification
The model specification being used in this example is the following:
$$
\begin{aligned}
V_{i, \textrm{Train}} &= \textrm{ASC Train} + \\
&\quad \beta _{ \textrm{tt_transit} } \textrm{Travel Time} _{ \textrm{Train}} * \frac{1}{60} + \\
&\quad \beta _{ \textrm{tc_train} } \textrm{Travel Cost}_{\textrm{Train}} * \left( GA == 0 \right) * 0.01 + \\
&\quad \beta _{ \textrm{headway_train} } \textrm{Headway} _{\textrm{Train}} * \frac{1}{60} + \\
&\quad \beta _{ \textrm{survey} } \left( \textrm{Train Survey} == 1 \right) \\
\\
V_{i, \textrm{Swissmetro}} &= \textrm{ASC Swissmetro} + \\
&\quad \beta _{ \textrm{tt_transit} } \textrm{Travel Time} _{ \textrm{Swissmetro}} * \frac{1}{60} + \\
&\quad \beta _{ \textrm{tc_sm} } \textrm{Travel Cost}_{\textrm{Swissmetro}} * \left( GA == 0 \right) * 0.01 + \\
&\quad \beta _{ \textrm{headway_sm} } \textrm{Heaway} _{\textrm{Swissmetro}} * \frac{1}{60} + \\
&\quad \beta _{ \textrm{seat} } \left( \textrm{Seat Configuration} == 1 \right) \\
&\quad \beta _{ \textrm{survey} } \left( \textrm{Train Survey} == 1 \right) \\
&\quad \beta _{ \textrm{first_class} } \left( \textrm{First Class} == 0 \right) \\
\\
V_{i, \textrm{Car}} &= \beta _{ \textrm{tt_car} } \textrm{Travel Time} _{ \textrm{Car}} * \frac{1}{60} + \\
&\quad \beta _{ \textrm{tc_car}} \textrm{Travel Cost}_{\textrm{Car}} * 0.01 + \\
&\quad \beta _{\textrm{luggage}=1} \left( \textrm{Luggage} == 1 \right) + \\
&\quad \beta _{\textrm{luggage}>1} \left( \textrm{Luggage} > 1 \right)
\end{aligned}
$$

# Scale the travel time column by 60 to convert raw units (minutes) to hours
sm_long["travel_time_hrs"] = sm_long["travel_time"] / 60.0
# Scale the headway column by 60 to convert raw units (minutes) to hours
sm_long["headway_hrs"] = sm_long["headway"] / 60.0
# Get free_ticket
sm_long["free_ticket"] = (((sm_long["GA"] == 1) | (sm_long["WHO"] == 2)) &
                            sm_long["mode_id"].isin([1,2])).astype(int)
# Scale travel cost to 100.0
sm_long["travel_cost_hundreth"] = (sm_long["travel_cost"] *
                                  (sm_long["free_ticket"] == 0) / 100.0)
# Create a dummy variable for whether a person has a single piece of luggage
sm_long["single_luggage_piece"] = (sm_long["LUGGAGE"] == 1).astype(int)
# Create a dummy variable for whether a person has multiple pieces of luggage
sm_long["multiple_luggage_pieces"] = (sm_long["LUGGAGE"] == 3).astype(int)
# Create a dummy variable indicating that a person is NOT first class
sm_long["regular_class"] = 1 - sm_long["FIRST"]
# Create a dummy variable indicating that the survey was taken aboard a train
sm_long["train_survey"] = 1 - sm_long["SURVEY"]

specification = OrderedDict()
specification["intercept"] = [1, 2]
specification["travel_time_hrs"] = [[1, 2,], 3]
specification["travel_cost_hundreth"] = [1, 2, 3]
specification["headway_hrs"] = [1, 2]
specification["seat_configuration"] = [2]
specification["train_survey"] = [[1, 2]]
specification["regular_class"] = [1]
specification["single_luggage_piece"] = [3]
specification["multiple_luggage_pieces"] = [3]

# Estimate the multinomial logit model (MNL)
swissmetro_mnl = pl.create_choice_model(data=sm_long,
                                        alt_id_col="mode_id",
                                        obs_id_col="custom_id",
                                        choice_col="CHOICE",
                                        specification=specification,
                                        model_type="MNL")

# Specify the initial values and method for the optimization.
swissmetro_mnl.fit_mle(np.zeros(14))

# Look at the estimation results
swissmetro_mnl.get_statsmodels_summary()

### Evaluation metrics

from sklearn.metrics import accuracy_score, f1_score, classification_report

def model_pred(data, model, alt_id_col, obs_id_col, choice_col):
    data['predicted'] = model.predict(data)  
    is_chosen = data.groupby([obs_id_col])['predicted'].idxmax()
    data['predicted_choice'] = 0
    data.loc[is_chosen.values,'predicted_choice'] = 1
    
    actual = data.loc[data[choice_column] ==1,alt_id_column]
    pred = data.loc[data['predicted_choice'] ==1,alt_id_column]
    return actual, pred   

def nll():
    pass

alt_id_column = "mode_id"
obs_id_column = "custom_id"
choice_column = "CHOICE"

##### MNL model results

actual, predict = model_pred(sm_long, swissmetro_mnl, alt_id_column, obs_id_column, choice_column)
print(classification_report(actual, predict))

acc = accuracy_score(actual, predict), 
f1 = f1_score(actual, predict, average="weighted")
print(f"Accuracy = {np.round(acc,5)}, F1_score = {np.round(f1,5)}")
