==> 01_statsmodels_MNL.ipynb <==
# default_exp statsmodels_MNL

# MNL with Statsmodels

> API details of statsmodels_MNL.

#hide
from nbdev.showdoc import *

#### MNL model

#export
import numpy as np
import statsmodels.api as sm
import pandas as pd
import patsy

# Load data
anes_data = sm.datasets.anes96.load(as_pandas=False)
anes_exog = anes_data.exog
anes_exog = sm.add_constant(anes_exog, prepend=False)
print(anes_data.exog[:5,:])
print(anes_data.endog[:5])

# Fit MNL model
mlogit_mod = sm.MNLogit(anes_data.endog, anes_exog)
mlogit_res = mlogit_mod.fit()
print(mlogit_res.params)

mlogit_res.summary()
