==> 02_MNL_swissmetro.ipynb <==
# default_exp MNL_swissmetro

# MNL with Swissmetro dataset => Chang Train same as TasteNet

> API details of MNL with Swissmetro dataset.

#hide
from nbdev.showdoc import *

#### Data preparation

#export
import pickle
import numpy as np
import pandas as pd
import biogeme.database as db
import biogeme.biogeme as bio
import biogeme.models as models
from biogeme.expressions import Beta

# Convert .dat to .pck
df = pd.read_csv('./data/swissmetro.dat', sep='	')
# Remove unknown age (6)
df['AGE'].value_counts
df = df[df['AGE'] != 6]
# Remove "other" trip purpose (9)
df = df[df['PURPOSE'] != 9]
# Remove "unknown" choice (0)
df = df[df['CHOICE'] != 0]
df.shape

# Save to pickle file
df.to_pickle('./data/swissmetro_clean.pkl')

#### Data loading

data = pickle.load(open('./data/swissmetro_clean.pkl', "rb"))
data.head()

df.describe()

### Biogeme

# Read the data
# df = pd.read_csv('swissmetro.dat', '\t')
df = pickle.load(open('./data/swissmetro_clean.pkl', "rb"))
database = db.Database('swissmetro', df)

df.shape

# The following statement allows you to use the names of the variable
# as Python variable.
globals().update(database.variables)

# Removing some observations
# exclude = ((PURPOSE != 1) * (PURPOSE != 3) + (CHOICE == 0)) > 0
# database.remove(exclude)

# Parameters to be estimated
# ASC_CAR = Beta('ASC_CAR', 0, None, None, 0)
# ASC_TRAIN = Beta('ASC_TRAIN', 0, None, None, 0)
# ASC_SM = Beta('ASC_SM', 0, None, None, 1)
# B_TIME = Beta('B_TIME', 0, None, None, 0)
# B_COST = Beta('B_COST', 0, None, None, 0)


ASC_TRAIN = Beta('ASC_TRAIN', 0, None, None, 0)
ASC_SM = Beta('ASC_SM', 0, None, None, 0)
ASC_CAR = Beta('ASC_CAR',0, None, None, 1) # Constant

B_TIME_TRAIN = Beta('B_TIME_TRAIN', 0, None, None, 0)
B_TIME_SM = Beta('B_TIME_SM', 0, None, None, 0)
B_TIME_CAR = Beta('B_TIME_CAR', 0, None, None, 0)

B_HEADWAY_TRAIN = Beta('B_HEADWAY_TRAIN', 0, None, None, 0)
B_HEADWAY_SM = Beta('B_HEADWAY_SM', 0, None, None, 0)


B_COST_TRAIN = Beta('B_COST_TRAIN', -1, None, None, 1) # Constant
B_COST_SM = Beta('B_COST_SM', -1, None, None, 1) # Constant
B_COST_CAR = Beta('B_COST_CAR', -1, None, None, 1) # Constant



df.head().T

# Definition of new variables
SM_COST = SM_CO * (GA == 0)
TRAIN_COST = TRAIN_CO * (GA == 0)

CAR_AV_SP = CAR_AV * (SP != 0)
TRAIN_AV_SP = TRAIN_AV * (SP != 0)

TRAIN_TT_SCALED = TRAIN_TT # / 100.0
TRAIN_COST_SCALED = TRAIN_COST # / 100

SM_TT_SCALED = SM_TT # / 100.0
SM_COST_SCALED = SM_COST # / 100

CAR_TT_SCALED = CAR_TT # / 100
CAR_CO_SCALED = CAR_CO # / 100

TRAIN_HEADWAY = TRAIN_HE
SM_HEADWAY = SM_HE

# Definition of the utility functions
# V1 = ASC_TRAIN + \
#      B_TIME * TRAIN_TT_SCALED + \
#      B_COST * TRAIN_COST_SCALED
# V2 = ASC_SM + \
#      B_TIME * SM_TT_SCALED + \
#      B_COST * SM_COST_SCALED
# V3 = ASC_CAR + \
#      B_TIME * CAR_TT_SCALED + \
#      B_COST * CAR_CO_SCALED

V1 = ASC_TRAIN + B_TIME_TRAIN * TRAIN_TT_SCALED + B_HEADWAY_TRAIN * TRAIN_HEADWAY +  B_COST_TRAIN * TRAIN_COST_SCALED
V2 = ASC_SM + B_TIME_SM * SM_TT_SCALED + B_HEADWAY_SM * SM_HEADWAY + B_COST_SM * SM_COST_SCALED
V3 = ASC_CAR + B_TIME_CAR * CAR_TT_SCALED + B_COST_CAR * CAR_CO_SCALED

# Associate utility functions with the numbering of alternatives
V = {1: V1,
     2: V2,
     3: V3}

# Associate the availability conditions with the alternatives
av = {1: TRAIN_AV_SP,
      2: SM_AV,
      3: CAR_AV_SP}

# Definition of the model. This is the contribution of each
# observation to the log likelihood function.
logprob = models.loglogit(V, av, CHOICE)

# Create the Biogeme object
biogeme = bio.BIOGEME(database, logprob)
# biogeme.modelName = '01logit'
biogeme.modelName = 'anything'

# Estimate the parameters
results = biogeme.estimate()

# Get the results in a pandas table
pandasResults = results.getEstimatedParameters()
print(pandasResults)

pandasResults['Value']

pandasResults.at['ASC_CAR','Value']

# The validation consists in organizing the data into several slices
# of about the same size, randomly defined. Each slice is considered
# as a validation dataset. The model is then re-estimated using all
# the data except the slice, and the estimated model is applied on the
# validation set (i.e. the slice). The value of the log likelihood for
# each observation in the validation set is reported in a
# dataframe. As this is done for each slice, the output is a list of
# dataframes, each corresponding to one of these exercises.

validation_results = biogeme.validate(results,5)

for slide in validation_results:
    print(f'Log likelihood for {slide.shape[0]} validation data: {slide["Loglikelihood"].sum()}')
    

slices = 5

validationData = database.split(slices)

type(validationData)

for v in validationData:
    print(type(v))

type(v), len(v)

type(v[0]), type(v[1]), v[0].shape, v[1].shape

keepDatabase = database
type(keepDatabase)

estimationResults = results

allSimulationResults = []
for v in validationData:
    # v[0] is the estimation data set
    database = db.Database('Estimation data', v[0])
    biogeme.loglike.changeInitValues(estimationResults.getBetaValues())
    results = biogeme.estimate()
    simulate = {'Loglikelihood': biogeme.loglike}
#     simBiogeme = BIOGEME(db.Database('Validation data', v[1]),
#                          simulate)
#     simResult = simBiogeme.simulate(results.getBetaValues())
    simResult = biogeme.simulate(results.getBetaValues())
    allSimulationResults.append(simResult)
biogeme.database = keepDatabase

allSimulationResults

# Get accuracy
type(slide), slide.shape

results

### PyLogit

#export
from collections import OrderedDict
import pandas as pd
import numpy as np
import pylogit as pl

wide_sm = pd.read_table('./data/swissmetro.dat', sep="\t")


wide_sm.shape

include_criteria = (wide_sm.PURPOSE.isin([1, 3]) & 
                   (wide_sm.CHOICE !=0))
wide_sm = wide_sm.loc[include_criteria].copy()

wide_sm.shape

wide_sm.head()

#### Convert Wide to Long format

wide_sm.ID.nunique(), wide_sm.shape[0]/wide_sm.ID.nunique()

wide_sm[wide_sm.ID.isin([2])].head(10).T

wide_sm.columns

# Select and group suitable type of varialbes
# Create the list of individual specific variables
ind_variables = wide_sm.columns.tolist()[:15]
availability_variables = {1: 'TRAIN_AV',
                          2: 'SM_AV',
                          3: 'CAR_AV'}
# Alternative variables
alt_varying_variables = {u'travel_time': dict([(1, 'TRAIN_TT'),
                                               (2, 'SM_TT'),
                                               (3, 'CAR_TT')]),
                         u'travel_cost': dict([(1, 'TRAIN_CO'),
                                               (2, 'SM_CO'),
                                               (3, 'CAR_CO')]),
                         u'headway': dict([(1, 'TRAIN_HE'),
                                           (2, 'SM_HE')]),
                         u'seat_configuration': dict([(2,"SM_SEATS")])}
alt_varying_variables

# Define custom alternative ids
custom_alt_id = "mode_id"
# Create a custon id column
obs_id_column = "custom_id"
wide_sm[obs_id_column] = np.arange(wide_sm.shape[0], dtype=int) + 1

# Choice column
choice_column = "CHOICE"

# Convert to long-format
long_sm = pl.convert_wide_to_long(wide_sm,
                                  ind_variables,
                                  alt_varying_variables,
                                  availability_variables,
                                  obs_id_column,
                                  choice_column,
                                  new_alt_id_name=custom_alt_id
                                  )

long_sm.CHOICE.value_counts(), long_sm.shape

long_sm.head(10).T

#### Variable transformations

# Scale minutes to hours
long_sm['travel_time_hrs'] = long_sm['travel_time']/60.0
long_sm['headway_hrs'] = long_sm['headway']/60.0

long_sm["free_ticket"] = (((long_sm["GA"] == 1) |
                            (long_sm["WHO"] == 2)) &
                           long_sm[custom_alt_id].isin([1,2])).astype(int)

long_sm["travel_cost_hundreth"] = (long_sm["travel_cost"] *
                                    (long_sm["free_ticket"] == 0) / 100.0)
# Dummy variables
long_sm['single_luggage_piece'] = (long_sm['LUGGAGE'] == 1).astype(int)
long_sm['multiple_luggage_pieces'] = (long_sm['LUGGAGE'] == 3).astype(int)
long_sm['regular_class'] = 1 - long_sm['FIRST']
long_sm['train_survey'] = 1 - long_sm['SURVEY']


long_sm.head().T

basic_specf = OrderedDict()
basic_names = OrderedDict()

# NOTE: - Specification and variable names must be ordered dictionaries.
#       - Keys should be variables within the long format dataframe.
#         The sole exception to this is the "intercept" key.
#       - For the specification dictionary, the values should be lists
#         of integers or lists of lists of integers. Within a list, 
#         or within the inner-most list, the integers should be the 
#         alternative ID's of the alternative whose utility specification 
#         the explanatory variable is entering. Lists of lists denote 
#         alternatives that will share a common coefficient for the variable
#         in question.

basic_specification = OrderedDict()
basic_names = OrderedDict()

basic_specification["intercept"] = [1, 2]
basic_names["intercept"] = ['ASC Train',
                            'ASC Swissmetro']

basic_specification["travel_time_hrs"] = [[1, 2,], 3]
basic_names["travel_time_hrs"] = ['Travel Time, units:hrs (Train and Swissmetro)',
                                  'Travel Time, units:hrs (Car)']

basic_specification["travel_cost_hundreth"] = [1, 2, 3]
basic_names["travel_cost_hundreth"] = ['Travel Cost * (Annual Pass == 0), units: 0.01 CHF (Train)',
                                       'Travel Cost * (Annual Pass == 0), units: 0.01 CHF (Swissmetro)',
                                       'Travel Cost, units: 0.01 CHF (Car)']

basic_specification["headway_hrs"] = [1, 2]
basic_names["headway_hrs"] = ["Headway, units:hrs, (Train)",
                              "Headway, units:hrs, (Swissmetro)"]

basic_specification["seat_configuration"] = [2]
basic_names["seat_configuration"] = ['Airline Seat Configuration, base=No (Swissmetro)']

basic_specification["train_survey"] = [[1, 2]]
basic_names["train_survey"] = ["Surveyed on a Train, base=No, (Train and Swissmetro)"]

basic_specification["regular_class"] = [1]
basic_names["regular_class"] = ["First Class == False, (Swissmetro)"]

basic_specification["single_luggage_piece"] = [3]
basic_names["single_luggage_piece"] = ["Number of Luggage Pieces == 1, (Car)"]

basic_specification["multiple_luggage_pieces"] = [3]
basic_names["multiple_luggage_pieces"] = ["Number of Luggage Pieces > 1, (Car)"]


basic_specification

basic_names

# Estimate the multinomial logit model (MNL)
swissmetro_mnl = pl.create_choice_model(data=long_sm,
                                        alt_id_col=custom_alt_id,
                                        obs_id_col=obs_id_column,
                                        choice_col=choice_column,
                                        specification=basic_specification,
                                        model_type="MNL",
                                        names=basic_names)

# Specify the initial values and method for the optimization.
swissmetro_mnl.fit_mle(np.zeros(14))

# Look at the estimation results
swissmetro_mnl.get_statsmodels_summary()

# Look at other all results at the same time
swissmetro_mnl.print_summaries()
