==> 21_dataset.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('./data/swissmetro_clean.csv')
df.shape

df.head()

df.columns

df.CHOICE.value_counts()

alt_cols = {'TRAIN' : ['TRAIN_TT', 'TRAIN_CO', 'TRAIN_HE', 'TRAIN_AV'],
            'SM': ['SM_TT', 'SM_CO', 'SM_HE', 'SM_AV'],
            'CAR' : ['CAR_TT', 'CAR_CO', 'CAR_AV']
           }
alt_cols

def create_col_index(col_dict):
    h1 = []
    h2 = []
    for k,v in alt_cols.items():        
        for i in range(len(v)): h1.append(k)
        h2.extend(v)
    return [h1, h2]

selected_cols = create_col_index(alt_cols)

df[selected_cols[1]]

test = df[selected_cols[1]].copy()

test.shape

test.columns = selected_cols

test.head()

test.columns

test['TRAIN']
