==> 80_london_mode_choice.ipynb <==
import pandas as pd

df = pd.read_csv('./data/lpmc.dat', sep='\t')
df.shape

df.head()

df.columns

df.travel_mode.value_counts()

df.purpose.value_counts()

df.head()

df.shape

df.household_id.value_counts()
