==> 90_indexing.ipynb <==
import numpy as np
import pandas as pd

### Pandas index

obj = pd.Series(range(3), index=['a', 'b', 'c'])
obj

obj.index

labels = pd.Index(np.arange(3))
labels

obj2 = pd.Series([1.5, -2.5, 0.5], index=labels)COLO
obj2

dup_labels = pd.Index(['foo', 'foo', 'bar', 'bar'])
dup_labels

### Reindexing

obj = pd.Series([4.5, 7.3, -7.4, 2.], index=['d', 'b', 'a', 'c'])
obj

obj2 = obj.reindex(['a', 'b', 'c', 'd', 'e'])
obj2

### Renaming axis indexes

data = pd.DataFrame(np.arange(12).reshape((3,4)), 
                    index=['Ohio', 'Colorado', 'New York'],
                    columns=['one', 'two', 'three', 'four'])
data

transform = lambda x : x[:4].upper()
data.index.map(transform)

data.index = data.index.map(transform)
data

data

# Option 2 is used renameCOLO
data.rename(index={'NEW ':'NY'}, columns={'four':'FO'}, inplace=True)
data

### Hierarchical indexing

data = pd.Series(np.random.randn(9),
                index=[['a', 'a', 'a', 'b', 'b', 'c', 'c', 'd', 'd'], 
                       [1, 2, 3, 1, 3, 1, 2, 2, 3]])
data

data.index

data['a']

data['b':'c']

data.loc[:, 2]

data.unstack()

data.unstack().stack()

### Hierarchical columns

df = pd.DataFrame(np.arange(12).reshape((4,3)),
                  index=[['a','a','b','b'], [1, 2, 1, 2]],
                  columns=[['Ohio', 'Ohio', 'Colorado'],
                           ['Green', 'Red', 'Green']])
df

df['Ohio']

df['Ohio'].Green.values

df.index

df.columns

##### Names of index and columns

df

df.index.names = ['key 1', 'key 2']
df.columns.names = ['state', 'color']
df
